refactor(ai-service): Step 8 — consolidate skills into app/skills/
The skill registry, skill-prompt formatters, and label-catalog helper
all converge on app/skills/. Yamls live FLAT in the same directory
as the registry code (no nested skills/ subdir, no per-skill folder).
This matches the existing rule that skill yamls are data-only and
sit at app/skills/<name>.yaml.

Moves:
  app/skill_manager/__init__.py    →  app/skills/__init__.py
  app/skill_manager/_embedder.py   →  app/skills/_embedder.py
  app/skill_manager/_query.py      →  app/skills/_query.py
  app/skill_manager/_registry.py   →  app/skills/_registry.py
  app/skill_manager/skills/graph_analysis.yaml     →  app/skills/graph_analysis.yaml
  app/skill_manager/skills/lap_annotation.yaml     →  app/skills/lap_annotation.yaml
  app/skill_manager/skills/sub_label_catalog.yaml  →  app/skills/sub_label_catalog.yaml
  app/services/llm/skill_prompts.py  →  app/skills/prompts.py
  app/services/llm/label_catalog.py  →  app/skills/label_catalog.py

The old app/skill_manager/ package is deleted entirely.

Registry changes:
  • _SKILLS_ROOT was _PACKAGE_ROOT/"skills" (nested) — now _PACKAGE_ROOT
    (flat). The glob "*.yaml" plus the leading-underscore filter already
    skips the registry's own _embedder/_query/_registry python files.
  • docstring updated to describe the flat layout under app/skills/.

Importers updated (12 files): all references rewritten from
`app.skill_manager*` → `app.skills*`, plus
`app.services.llm.skill_prompts` → `app.skills.prompts`, and
`app.services.llm.label_catalog` → `app.skills.label_catalog`.

Validated: SkillRegistry loads all 3 yamls (graph_analysis,
lap_annotation, sub_label_catalog); label_catalog merge of
LABEL_MAPPING + sub_label yaml still works; /health 200.
