+++ acla_ai_service/app/services/llm/agent/sub_agents/label_verifier.py
@@ -20,8 +20,8 @@ import logging
 from typing import Any, Dict, List, Tuple
 
 from app.domain.labels import LABEL_MAPPING
-from app.skill_manager import embed
-from app.services.llm.label_catalog import find_labels, get_label
+from app.skills import embed
+from app.skills.label_catalog import find_labels, get_label
 from app.services.llm.agent.framework import Agent, AgentState
 from app.services.llm.agent.evaluators import (
     AttachmentPool,
