+++ acla_ai_service/app/services/llm/annotation/followup.py
@@ -29,8 +29,8 @@ from dataclasses import dataclass, field
 from typing import Any, Callable, Dict, List, Optional
 
 from app.domain.labels import LABEL_MAPPING
-from app.services.llm.label_catalog import find_labels, get_label
-from app.services.llm.skill_prompts import graph_analysis_prompt
+from app.skills.label_catalog import find_labels, get_label
+from app.skills.prompts import graph_analysis_prompt
 
 LOGGER = logging.getLogger(__name__)
