+++ acla_ai_service/app/services/llm/annotation/tools/__init__.py
@@ -49,7 +49,7 @@ def list_eligible_labels(
         }
     """
     from app.services.llm.agent.evaluators import PipelineAttachment
-    from app.services.llm.label_catalog import find_labels, get_label
+    from app.skills.label_catalog import find_labels, get_label
 
     def _attach(groups: List[Dict[str, Any]]) -> "PipelineAttachment":
         return PipelineAttachment(
