+++ acla_ai_service/app/skills/_registry.py
@@ -2,9 +2,9 @@
 
 Layout::
 
-    app/skill_manager/                 (this package — orchestrator code)
-      __init__.py, _registry.py, _query.py, _embedder.py
-      skills/<name>.yaml               (pure data — drop a yaml in, restart)
+    app/skills/                        (this package — orchestrator code + yaml data)
+      __init__.py, _registry.py, _query.py, _embedder.py    (code)
+      <name>.yaml                      (skill definitions — drop a yaml in, restart)
 
 A skill is a single yaml file. No Python escape hatches — filtering,
 formatting, or merging with non-skill state belong to the caller.
@@ -33,7 +33,7 @@ from typing import Any, Dict, List, Optional
 import numpy as np
 import yaml
 
-from app.skill_manager._query import (
+from app.skills._query import (
     _split_path,
     find as _filter,
     get_path,
@@ -43,7 +43,9 @@ from app.skill_manager._query import (
 LOGGER = logging.getLogger(__name__)
 
 _PACKAGE_ROOT = Path(__file__).resolve().parent
-_SKILLS_ROOT = _PACKAGE_ROOT / "skills"
+# Yamls live FLAT under app/skills/<name>.yaml (no subdir) — see
+# refactor/hexagonal-v1 Step 8 and feedback_skill_is_data_only.md.
+_SKILLS_ROOT = _PACKAGE_ROOT
 _CACHE_DIR = _PACKAGE_ROOT.parent.parent / ".cache" / "skills"
 _EMBEDDINGS_FILE = _CACHE_DIR / "embeddings.npz"
 _MANIFEST_FILE = _CACHE_DIR / "manifest.json"
@@ -166,7 +168,7 @@ class SkillRegistry:
                 to_embed.append(name)
 
         if to_embed:
-            from app.skill_manager._embedder import embed
+            from app.skills._embedder import embed
             texts = [discovery_texts[n] for n in to_embed]
             vecs = embed(texts)
             if vecs.ndim == 1:
@@ -256,7 +258,7 @@ class SkillRegistry:
         """Embedding-similarity search over discovery headers."""
         if self._embeddings is None or not self._names:
             return []
-        from app.skill_manager._embedder import embed
+        from app.skills._embedder import embed
         q = embed(query)
         scores = self._embeddings @ q
         order = np.argsort(-scores)[: max(0, top_k)]
