+++ acla_ai_service/app/skills/lap_annotation.yaml
